Preliminary Harmonize Adaptation using Inlining
The inline-based adaptation has thusfar not worked due to instruction re-ordering during compilation. Nonethless, inlining may be fixed in the future, and this commit represents the version that is closest to working.

Aside from inlining-specific logic, most of the remaining code is used in the future (working) version that uses linking instead.